# Vectorized HttpCharacters