… src/Shared/ServerInfrastructure/HttpCharacters.cs