@gfoidl 2022-10-20
I'll close this PR, as I'm pretty sure with https://github.com/dotnet/runtime/issues/68328 we get a proper building block and the wire-up should become trivial.

Only the "extended case" isn't covered by the runtime-issue. In the worst case it remains the scalar-approach as it's now, so no regression should happen, but also no improvement.