@gfoidl 2022-11-22
https://github.com/dotnet/runtime/issues/68328 is done 🎉, so in the next days I'll create a PR based on these new APIs.